% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{against_numbers}
\alias{against_numbers}
\title{against_numbers}
\usage{
against_numbers(n1 = 1:5, n2 = 1:100)
}
\arguments{
\item{n1}{is  vector of integer sequence number}

\item{n2}{is  vector of integer sequence number}
}
\value{
This function retunr a vector n2 eliminate n1
}
\description{
against_numbers
}
\examples{
against_numbers(1:5,1:100)
}
