% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{area}
\alias{area}
\title{area return a  area from polygon xy coordinates}
\usage{
area(xy)
}
\arguments{
\item{xy}{a atrix or data frame of xy coordinates}
}
\value{
a vector of polygon area
}
\description{
area return a  area from polygon xy coordinates
}
\examples{
xy<-data.frame(x=c(0,1,2,1),y=c(1,2,1,0))
plot(xy,type="b") ; polygon(xy)
area(xy)

}
