% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRingGrowth.R
\name{cor2yr}
\alias{cor2yr}
\title{Title}
\usage{
cor2yr(idx, met, y1 = 1901, y2 = 1950)
}
\arguments{
\item{y2}{}
}
\description{
Title
}
\examples{
plot(rownames(met),met[,13],type="b")
plot(ya,type="b")
idx<-TreeRingIndex(ya)$idx
 r12<-cor2yr(idx,met,y1=1901,y2=1955)
plot(r12,type="h",lwd=15)
bp<-barplot(r12)
abline(h=c(0),lty=1,lwd=3,col="red")
abline(v=c(mean(bp[12:13])),lty=2,lwd=3)
text(bp,r12*0.9,c(1:12,1:12))
}
