% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{Cosine_similarity}
\alias{Cosine_similarity}
\title{return a cosine similarity of two vector}
\usage{
Cosine_similarity(g1, g2)
}
\arguments{
\item{g1}{numeric vector}

\item{g2}{numeric vector}
}
\value{
a cosine similarity of two vector
}
\description{
return a cosine similarity of two vector
}
\examples{
g1=1:10 ; g2=1:10#-1
plot(g1,type="b");lines(g2,type="b",col="red")
Cosine_similarity(g1,g2)

g1=1:10 ; g2=c(2,6,4,3,1,6,7,9,8,11)
plot(g1,type="b");lines(g2,type="b",col="red")
#' Cosine_similarity(g1,g2)

 g1=1:10 ; g2=10:1
plot(g1,type="b");lines(g2,type="b",col="red")
Cosine_similarity(g1,g2)

g1=1:10 ; g2=runif(10)*10
plot(g1,type="b");lines(g2,type="b",col="red")
Cosine_similarity(g1,g2)

}
