% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiameterGrowthByArea.R
\name{DiameterGrowthByArea}
\alias{DiameterGrowthByArea}
\title{return data frame and plot of tree trunk diameter estimated by annual ring area}
\usage{
DiameterGrowthByArea(ya, ...)
}
\arguments{
\item{ya}{data frame of Year and TreeRingArea}
}
\value{
data frame of Year and diameter_cm
}
\description{
return data frame and plot of tree trunk diameter estimated by annual ring area
}
\examples{
DiameterGrowthByArea(ya)

}
