% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{dst}
\alias{dst}
\title{dst return a vector of distances from original a point (0,0)
from a matrix or data frame of xy coordinates}
\usage{
dst(xy)
}
\arguments{
\item{xy}{a matrix or data frame of xy coordinates}
}
\value{
a vector of distances from original a point
}
\description{
dst return a vector of distances from original a point (0,0)
from a matrix or data frame of xy coordinates
}
\examples{
plot(L[[1]])
plot(dst(L[[1]]))
}
