% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{dstpp}
\alias{dstpp}
\title{dstpp returns vector distance between adjacent two points}
\usage{
dstpp(x, y)
}
\arguments{
\item{x}{vector of x coordinates}

\item{y}{vector of y coordinates}
}
\value{
vector distance between adjacent two points
}
\description{
dstpp returns vector distance between adjacent two points
}
\examples{
x<-l.[,1] ;y<-l.[,2]
dstpp(x,y)
}
