% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRingGrowth.R
\name{Ga_w}
\alias{Ga_w}
\title{estimating growth of trunk area from tree ring width (trw)}
\usage{
Ga_w(trw, id)
}
\arguments{
\item{trw}{list of tree ring width (cambium to pith)}

\item{id}{numeric, id of radial annual ring points}
}
\value{
numeric vector of trunk cross-sectional area
}
\description{
estimating growth of trunk area from tree ring width (trw)
}
