% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncompleteTreeRingFix.R
\name{IncompleteTreeRings_MissingRings}
\alias{IncompleteTreeRings_MissingRings}
\title{return reference number of tree ring polygons (L) with missing range determined median magnification}
\usage{
IncompleteTreeRings_MissingRings(L, mag.med = 30)
}
\arguments{
\item{mag.med}{median magnification, threshold for decision as missing ring range}
}
\value{
numeric vector of L list reference number
}
\description{
return reference number of tree ring polygons (L) with missing range determined median magnification
}
\examples{
i<-IncompleteTreeRings_MissingRings(L,mag.med=200)
length(i)
Lplot2(L,i)

}
