% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncompleteTreeRingFix.R
\name{IncompleteTreeRingsBind}
\alias{IncompleteTreeRingsBind}
\title{IncompleteTreeRingsBind bind complete tree ring line to incomplete tree ring part}
\usage{
IncompleteTreeRingsBind(L, rn = 2, rn0 = 1, plot.yn = "FALSE")
}
\arguments{
\item{rn}{reference number (tree ring list ex. L) of   incomplete tree ring}

\item{rn0}{reference number (tree ring list ex. L) of adjacent complete tree ring}
}
\value{
data frame of x y coordinates compensated incomplete tree ring part
}
\description{
IncompleteTreeRingsBind bind complete tree ring line to incomplete tree ring part
}
\examples{

#load("L_000.RData");rn=2;rn0=1 #
#rn=3;rn0=2 # rn=4;rn0=3 # rn=17;rn0=18

load("L_001.RData")
rn=20;rn0=19
par(mfrow=c(1,3)) # par(mfrow=c(1,1))#'
plot(L[[rn]],type="b")
plot(L[[rn0]],type="b")
lnew<-IncompleteTreeRingsBind(L,rn,rn0,plot.yn=T)
plot(lnew,type="l",col="red")
par(mfrow=c(1,1))
plot(L[[rn0]],type="l") ; lines(lnew,type="l",col="red")


#
iTR<-IncompleteTreeRingsFix_setClass(L_filename="line277_h600_2",L_id.tag="",L_ring.tag="id")
iTR<-IncompleteTreeRingsFix_RingList(iTR)
P00<-ReadShapefile_P00("points277_h600",id.tag="id",ring.tag="ring")
iTR@P00 <-P00
L  <-Lsort_all( Lmove(iTR@L,-P00))
Lplot(L)                        # plot tree lings
L_ <- Llist2dataframe(L)
# manual correct
slotNames(iTR)
d<-iTR@inc.ring
idst<-c() ; irn<-c() ; n<-c()
for(i in iTR@yr[-1]){
n<-c(n,sum(L_$yr==i))
 idst<-c(idst,list(dst(d[d$yr==i,c("x","y")])))
 irn<-c(irn,list(d$pn[d$yr==i]))
}

names(irn) <- names(idst) <- iTR@yr[-1]
n<-data.frame(yr=iTR@yr[-1],n)
idst ; irn ;n
if(0){
i<-d[,c("yr1","yr2","yr3")]==2
d[,c("pn1","pn2","pn3")][i]
}

#edit(iTR@inc.ring)
iTR@pn  # 年輪合流点確認　(Qgisのsnap)
L2<-L  #　年輪修正用のコピー
(d <- iTR@inc_edit)　#作業行程ファイル　不完全年輪と参照年輪
(d<-d[-1,])  # outer ring omited 形成層は外す
#edit(d)
8年は9年を参照している　9年も不完全年輪で7年を参照している
(d<-d[c(1:7,9,8,11:nrow(d)),]) # 8年9年交代
d$rn.ref[which(d$yr==9)]<-8 #9年は7年 rn=8 を参照する
d$rn.ref[which(d$yr==8)]<-10 #8年は9年 rn=10を参照する



# edit(d)
#######　修復作業　>>>>

L2<-L
Lplot2(L2,d$rn)
#'
L2<-L
i12<-1:nrow(d) #;i12<-1:9
for( i in i12){#nrow(d)
j<-d$rn[i]
  L2[[j]]<-IncompleteTreeRingsBind(L2,j,d$rn.ref[i])
}
# <<<<<<<<<
Lplot2(L2,d$rn[i12],nrow=3,ncol=3,type="b",col="red")#'
#WriteShapefile_AnnualRings (Lmove(L2,P00), "test20221112")

}
