% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncompleteTreeRingFix.R
\name{IncompleteTreeRingsFix_RingList}
\alias{IncompleteTreeRingsFix_RingList}
\title{return data frame of points list duplicated,}
\usage{
IncompleteTreeRingsFix_RingList(iTR)
}
\arguments{
\item{iTR}{}
}
\value{
iTR
}
\description{
return data frame of points list duplicated,
}
\examples{
iTR<-IncompleteTreeRingsFix_RingList(iTR)
iTR@pn
}
