% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncompleteTreeRingFix.R
\name{IncompleteTreeRingsFix_setClass}
\alias{IncompleteTreeRingsFix_setClass}
\title{sets class for IncompleteTreeRingsFix, should run this function before start IncompleteTreeRingsFix}
\usage{
IncompleteTreeRingsFix_setClass(
  L_filename = "line277_h600_2",
  L_id.tag = "id",
  L_ring.tag = "ring"
)
}
\arguments{
\item{L_ring.tag}{}
}
\value{
class　of IncompleteTreeRingsFix
}
\description{
sets class for IncompleteTreeRingsFix, should run this function before start IncompleteTreeRingsFix
}
\examples{
iTR<-IncompleteTreeRingsFix_setClass(L_filename="line277_h600_2",L_id.tag="",L_ring.tag="id")
iTR<-IncompleteTreeRingsFix_RingList(iTR)
P00<-ReadShapefile_P00("points277_h600",id.tag="id",ring.tag="ring")
iTR@P00 <-P00
Lsort_all( Lmove(iTR@L,-P00))
}
