% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{Ldeg360}
\alias{Ldeg360}
\title{Ldeg360 return a vector of center angle 0 to 360(degree) for x y coordinate vector}
\usage{
Ldeg360(x, y)
}
\arguments{
\item{x}{is a vector of x coordinates}

\item{y}{is a vector of y coordinates}
}
\value{
a vector of center angle 0 to 360(degree) for x y coordinate vector
}
\description{
Ldeg360 return a vector of center angle 0 to 360(degree) for x y coordinate vector
}
\examples{
 xy <- L[[1]]
 Ldeg360(xy[,1],xy[2])
}
