% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{Llist2dataframe}
\alias{Llist2dataframe}
\title{Llist2dataframe  convert from a list of tree rings polygons (L)
to a data frame with no.,year,x,y,r(radius),radian(center angle),degree.
The data frame is sorted by degree(0 to 360).}
\usage{
Llist2dataframe(L)
}
\arguments{
\item{L}{}
}
\value{
data fraeme
}
\description{
Llist2dataframe  convert from a list of tree rings polygons (L)
to a data frame with no.,year,x,y,r(radius),radian(center angle),degree.
The data frame is sorted by degree(0 to 360).
}
\examples{
L_ <- Llist2dataframe(L)
head(L_) ; tail(L_)
}
