% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Lplot}
\alias{Lplot}
\title{Lplot plots a graphics of tree rings}
\usage{
Lplot(L, rn = 1:length(L), ...)
}
\arguments{
\item{L}{is a list of tree rings polygon coordinates (X,Y)}

\item{rn}{vector of ring number of list (L), default 1:length(L)}
}
\description{
Lplot plots a graphics of tree rings
}
\examples{
Lplot(L)
Lplot(L,rn=1:50,type="l")
}
