% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Lplot2}
\alias{Lplot2}
\title{Lplot2 plots a graphics of tree rings by 1 ring (3*3 in a screen)}
\usage{
Lplot2(L, i.ring = 1:length(L), nrow = 3, ncol = 3, ask = "FALSE", ...)
}
\arguments{
\item{L}{is a list of tree rings polygon coordinates (X,Y)}

\item{ncol}{}
}
\description{
Lplot2 plots a graphics of tree rings by 1 ring (3*3 in a screen)
}
\examples{
load("L.RData")
Lplot2(L,type="b")
load("L_001.RData")
Lplot2(L,type="b")
Lplot2(L,i.ring=1:9, type="b")
Lplot2(L,i.ring=1:9, nrow=1,ncol=1,type="b")
}
