% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Lrad.plot}
\alias{Lrad.plot}
\title{Lrad.plot check center angle of points to input order}
\usage{
Lrad.plot(L, i.ring = 1:length(L))
}
\arguments{
\item{i.ring}{}
}
\description{
Lrad.plot check center angle of points to input order
}
\examples{
load("L.RData")
Lrad.plot(L,i.ring)
load("L_001.RData")
Lrad.plot(L,i.ring)

}
