% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{Lrn}
\alias{Lrn}
\title{return a ring number of tree ring polygons list (L) from year}
\usage{
Lrn(L, yr)
}
\arguments{
\item{L}{tree ring polygons list (L)}

\item{yr}{years (or rings)}
}
\value{
a ring number of tree ring polygons list (L)
}
\description{
return a ring number of tree ring polygons list (L) from year
}
\examples{
Lrn(L,8)
}
