% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{Lsort}
\alias{Lsort}
\title{Lsort}
\usage{
Lsort(l.)
}
\arguments{
\item{l.}{x,y coordinates matrix (ncol=2) or data.frame of an annual ring.}
}
\value{
ordered with center angle of each point
}
\description{
Lsort
}
\examples{
save(l.,file="l..RData")
plot(l.,type="l")
plot(Lsort(l.),type="l")

}
