% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{Lsort_all}
\alias{Lsort_all}
\title{Lsort_all}
\usage{
Lsort_all(L)
}
\arguments{
\item{L}{a list of annual ring polygons (x,y)}
}
\value{
a list of annual ring polygons (x,y) ordered with center angle of each point
}
\description{
Lsort_all
}
\examples{
Lsort_all(L)
}
