% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncompleteTreeRingFix.R
\name{MissingRingCompensation}
\alias{MissingRingCompensation}
\title{MissingRingCompensation compensate lack part using with spline curve estimation of
relationships between center angles and radius}
\usage{
MissingRingCompensation(l., plot.yn = "TRUE")
}
\arguments{
\item{l.}{a matrix or data frame of x y coordinates with lack part of a tree ring}
}
\value{
a matrix  of x y coordinates that compensate lack part
}
\description{
MissingRingCompensation compensate lack part using with spline curve estimation of
relationships between center angles and radius
}
\examples{
par(mfrow=c(1,2))
plot(l.,type="b")
l.comp <- MissingRingCompensation(l.)
plot(l.comp,type="b")
L.comp<-L
L.comp[[1]]<-l.
Lplot(L.comp)
L.comp<-Lmove(L.comp,P00)
WriteShapefile_AnnualRings(L.comp,"L.comp")

}
