% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{nstP}
\alias{nstP}
\title{nstP return a vector of row numbers of  points that have nearest  center angle}
\usage{
nstP(z1, z2)
}
\arguments{
\item{z1}{a data frame or a matrix of  xy coordinates of a tree ring  (usualy inner ring)}

\item{z2}{a data frame or a matrix of  xy coordinates of a tree ring  (usualy outer ring)}
}
\value{
a vector of row numbers of z2, the length is nrow(z1)
}
\description{
nstP return a vector of row numbers of  points that have nearest  center angle
}
\examples{
L_out<-L[[1]];L_in<-L[[30]]
np<-nstP(L_out,L_in)
plot(L_out,col="red"); points(L_in)
segments(L_out[,1],L_out[,2],L_in[np,1],L_in[np,2],col="blue")

}
