% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_AnnualRing}
\alias{plot_AnnualRing}
\title{plot_AnnualRing
This function draws annual rings of a disk from x, y list(x,y) with name of year.}
\usage{
plot_AnnualRing(L = L, year = 0, ...)
}
\arguments{
\item{L}{list(x,y) of annual ring coordinates with name of year}

\item{year_label}{name of  column of annual ring year (0(cambium),1,2,....,n(pith))}
}
\description{
plot_AnnualRing
This function draws annual rings of a disk from x, y list(x,y) with name of year.
}
\examples{
plot_AnnualRing(L.,year=8)
plot_AnnualRing(L,year="0",type="b",col="red")
}
