% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_AnnualRing_df}
\alias{plot_AnnualRing_df}
\title{plot_AnnualRing_df
This function draws a annual ring of a disk from data frame(x,y,year).}
\usage{
plot_AnnualRing_df(df = L_, year = 0, year_label = "yr")
}
\arguments{
\item{df}{name of a data frame}

\item{year_label}{name of  column of annual ring year (0(cambium),1,2,....,n(pith))}
}
\description{
plot_AnnualRing_df
This function draws a annual ring of a disk from data frame(x,y,year).
}
\examples{
plot_AnnualRing_df(L_,1)

}
