% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_AnnualRings_df}
\alias{plot_AnnualRings_df}
\title{plot_AnnualRings_df
This function draws annual rings of a disk from data frame(x,y,year).}
\usage{
plot_AnnualRings_df(df = L_, year_label = "yr")
}
\arguments{
\item{df}{name of a data frame}

\item{year_label}{name of  column of annual ring year (0(cambium),1,2,....,n(pith))}
}
\description{
plot_AnnualRings_df
This function draws annual rings of a disk from data frame(x,y,year).
}
\examples{
plot_AnnualRings_df(L_)

}
