% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_year_RingArea}
\alias{plot_year_RingArea}
\title{plot and return data frame of  year_disk area and year_annual ring area}
\usage{
plot_year_RingArea(L2, yr_end = 2018)
}
\arguments{
\item{yr_end}{}
}
\description{
plot and return data frame of  year_disk area and year_annual ring area
}
\examples{
L2<-TR@L2
plot_year_RingArea(L2,2018)
}
