% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRingGrowth.R
\name{r2a}
\alias{r2a}
\title{estimate trank area from radius as true circle}
\usage{
r2a(r)
}
\arguments{
\item{r}{radius (length between a pith and cambium layer)}
}
\value{
numeric of  trunk area
}
\description{
estimate trank area from radius as true circle
}
