% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{rdst}
\alias{rdst}
\title{rdst return relative distance between two representative tree rings}
\usage{
rdst(L, P, yr)
}
\arguments{
\item{L_}{is data frame of tree ring}

\item{yr_out}{is a integer of year of outer ring}

\item{yr_in}{is a integer of year of inner ring}

\item{x}{is a value.  x coordinate of a target point}

\item{y}{is a value.  y coordinate of a target point}
}
\value{
is  a data frame with relative distance and center angle
}
\description{
rdst return relative distance between two representative tree rings
}
\examples{
rdst.<-rdst(L,P,73)
plot(rdst.)
spline<-smooth.spline(rdst.$rad,rdst.$rdst, spar =0.0002)
　lines(predict(spline,seq(-pi,pi,0.01)),col="red")
}
