% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{rdst_MerginePlus}
\alias{rdst_MerginePlus}
\title{rdst_MerginePlus return relative distance between two representative tree rings}
\usage{
rdst_MerginePlus(L, P, yr)
}
\arguments{
\item{L_}{is data frame of tree ring}

\item{yr_out}{is a integer of year of outer ring}

\item{yr_in}{is a integer of year of inner ring}

\item{x}{is a value.  x coordinate of a target point}

\item{y}{is a value.  y coordinate of a target point}
}
\value{
is  a data frame with relative distance and center angle(degree)
with mergine (-90 - 0 - 360 - 90)
}
\description{
rdst_MerginePlus return relative distance between two representative tree rings
}
\examples{
year.<-73
rdst.<-rdst_MerginePlus(L,P,year.)
plot(rdst.,xlim=c(-200,200),main=year.)
spline<-smooth.spline(rdst.$deg,rdst.$rdst, spar =0.0002)
　lines(predict(spline,seq(-202,220,1)),col="red")
　
　
}
