% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeFile.R
\name{ReadShapefile_AnnualRingPoints}
\alias{ReadShapefile_AnnualRingPoints}
\title{ReadShapefile_AnnualRingPoints}
\usage{
ReadShapefile_AnnualRingPoints(
  filename = "points277_h600",
  id.tag = "id",
  ring.tag = "ring"
)
}
\arguments{
\item{filename}{is a file name of annual ring points (shape file )
The extension (.shp) is unnecessary.}

\item{id}{string, column name of id (attribute table)

 @param ring string, column name of ring years  (0 is cambium layer)}
}
\value{
a data frame of AnnualRingPoints (radial input and correction points)
}
\description{
ReadShapefile_AnnualRingPoints
}
\examples{
#年輪計測点の読み込み ####
ReadShapefile_AnnualRingPoints("../155_h120/points155_h120",id.tag="id",ring.tag="ring")

}
