% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeFile.R
\name{ReadShapefile_Cracks}
\alias{ReadShapefile_Cracks}
\title{return a list of coordinates(x,y) of cracks from shape file}
\usage{
ReadShapefile_Cracks(filename = "loss277_h600_2", ring.tag = "year")
}
\arguments{
\item{filename}{shape file name of cracks}

\item{ring.tag}{column name of year}
}
\value{
a list of coordinates(x,y) of cracks
}
\description{
return a list of coordinates(x,y) of cracks from shape file
}
\examples{
ware <- ReadShapefile_Cracks(filename="loss277_h600_2",ring.tag="year")
Lplot(Lmove(TR@L2,TR@P00),col="gray")
sapply(ware,polygon,col="red")
sapply(ware,area)
plot(ya)
}
