% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeFile.R
\name{ReadShapefile_IncompleteTreeRingsPoints}
\alias{ReadShapefile_IncompleteTreeRingsPoints}
\title{return x,y coordinates of a tree ring center point (P00) from shape file of tree ring points}
\usage{
ReadShapefile_IncompleteTreeRingsPoints(filename, year_tag, year_ref.tag)
}
\arguments{
\item{filename}{is a file name of annual ring points (shape file )
The extension (.shp) is unnecessary.}

\item{id}{string, column name of id (attribute table)

 @param ring string, column name of ring years  (0 is cambium layer)}
}
\value{
numeric : x,y coordinates of a tree ring center point (P00)
}
\description{
return x,y coordinates of a tree ring center point (P00) from shape file of tree ring points
}
\examples{
#年輪計測点の読み込み ####
d<-ReadShapefile_IncompleteTreeRingsPoints("277_h600_IncompleteRingPoints_shp",year_tag="yr",year_ref.tag="yr.ref")
d

}
