% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeFile.R
\name{ReadShapefile_P00}
\alias{ReadShapefile_P00}
\title{return x,y coordinates of a tree ring center point (P00) from shape file of tree ring points}
\usage{
ReadShapefile_P00(
  filename = "points277_h600",
  id.tag = "id",
  ring.tag = "ring"
)
}
\arguments{
\item{filename}{is a file name of annual ring points (shape file )
The extension (.shp) is unnecessary.}

\item{id}{string, column name of id (attribute table)

 @param ring string, column name of ring years  (0 is cambium layer)}
}
\value{
numeric : x,y coordinates of a tree ring center point (P00)
}
\description{
return x,y coordinates of a tree ring center point (P00) from shape file of tree ring points
}
\examples{
#年輪計測点の読み込み ####
ReadShapefile_P00("points277_h600",id.tag="id",ring.tag="ring")

}
