% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{seq.deg}
\alias{seq.deg}
\title{seq.deg return vector of sequence of angles between start and end angle
 0 to pi -pi to 0}
\usage{
\method{seq}{deg}(deg1, deg2, deg.by = 1)
}
\arguments{
\item{deg1}{start angle}

\item{deg2}{end angle}

\item{deg.by}{step of sequence}
}
\value{
vector of sequence of angles between start and end angle
}
\description{
seq.deg return vector of sequence of angles between start and end angle
 0 to pi -pi to 0
}
\examples{
seq.deg(170,-170,.5)

}
