% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRingGrowth.R
\name{Takahashi2011_correlation_TreeRingIndex_Temperature}
\alias{Takahashi2011_correlation_TreeRingIndex_Temperature}
\title{plot barplot of Takahashi2011,correlations between tree ring index and monthly mean temperature}
\usage{
Takahashi2011_correlation_TreeRingIndex_Temperature()
}
\arguments{
\item{y2}{}
}
\value{
data frame of year and correlation  between tree ring index and monthly mean temperature
}
\description{
plot barplot of Takahashi2011,correlations between tree ring index and monthly mean temperature
}
\examples{
#data("Abies277_h600","Takayama_Temperature_1900-2018")
plot(ya,type="b")
idx <- TreeRingIndex(ya)$idx
plot(idx,type="b")
y1 = 1900 ; y2 = 1950
}
