% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRingGrowth.R
\name{TreeRingIndex}
\alias{TreeRingIndex}
\title{return tree ring index from sequence of year and tree ring area
using with spline curve}
\usage{
TreeRingIndex(ya)
}
\arguments{
\item{ya}{}
}
\value{
list of spline curve function("spline") and tree ring index("TreeRingIndex")
}
\description{
return tree ring index from sequence of year and tree ring area
using with spline curve
}
\examples{
par(mfrow=c(2,1))
plot(ya,type="b",ylab="Tree ring area",xlab="year")
data("Abies277_h600","")
sm <- TreeRingIndex(ya)$spline
idx<- TreeRingIndex(ya)$idx
lines(sm,col="red",lw=3)
plot(idx,type="b",ylab="Tree ring index",xlab="year")
abline(h=1,col="red")
abline(v=seq(1800,2020,10),col="gray")


}
