% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRingsInterpolation.R
\name{TreeRingsInterpolation}
\alias{TreeRingsInterpolation}
\title{TreeRingInterpolation}
\usage{
TreeRingsInterpolation(
  P_filename,
  L_filename,
  L2_filename,
  P_id.tag = "id",
  P_ring.tag = "ring",
  L_id.tag = "id",
  L_ring.tag = "ring"
)
}
\arguments{
\item{L_ring.tag}{}
}
\description{
TreeRingInterpolation
}
\examples{
TR<-TreeRingsInterpolation(
P_filename="points277_h600",
L_filename="277_h600不完全年輪追加",
L2_filename="Abies277_h600_interpolated",
P_id.tag="id",P_ring.tag="ring",
L_id.tag="id",L_ring.tag="ring")

str(L)
Lplot2(TR@L,1:length(TR@L))

#save(TR,file=paste0(TR@L2_filename,".RData"))
}
