% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRingsInterpolation.R
\name{TreeRingsInterpolation_calc}
\alias{TreeRingsInterpolation_calc}
\title{TreeRingsInterpolation_calc　interpolation tree ring between manual input tree rings with tree ring points}
\usage{
TreeRingsInterpolation_calc(TR)
}
\arguments{
\item{TR}{}
}
\value{
TR
}
\description{
TreeRingsInterpolation_calc　interpolation tree ring between manual input tree rings with tree ring points
}
\examples{
TreeRingsInterpolation_calc(TR)

}
