% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRingsInterpolation.R
\name{TreeRingsLines}
\alias{TreeRingsLines}
\title{TreeRingsLines}
\usage{
TreeRingsLines(TR)
}
\arguments{
\item{TR}{is tree ring  class (class_TreeRingsInterpolate )}
}
\description{
TreeRingsLines
}
\examples{
 TreeRingsLines(TR)
}
