% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRingsInterpolation.R
\name{TreeRingsPoints}
\alias{TreeRingsPoints}
\title{TreeRingsPoints read TreeRingsPoints shape file, check and save parameters}
\usage{
TreeRingsPoints(TR)
}
\arguments{
\item{TR}{is tree ring  class (class_TreeRingsInterpolate )

#@return list of  (P,P00,YR_P,n_id,YR_P,n_YR)}
}
\description{
TreeRingsPoints read TreeRingsPoints shape file, check and save parameters
}
\examples{
P_filename<-"points277_h600"		#####　<<<　各人シェープファイル名入力　(拡張子はいらない)
TreeRingsPoints(TR)

load(paste0(P_filename,".RData"))
}
