% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRingGrowth.R
\name{TreeRingWidthMatrix_ID}
\alias{TreeRingWidthMatrix_ID}
\title{return a numeric list of tree ring width from radial measurement points}
\usage{
TreeRingWidthMatrix_ID(P, idn = 8)
}
\arguments{
\item{P}{data frame of tree ring points}

\item{idn}{length of id of radial measurement points}
}
\value{
list of tree ring width
}
\description{
return a numeric list of tree ring width from radial measurement points
}
\examples{
trw<-TreeRingWidthMatrix_ID(P)
plot(trw[[1]],type="l")
for (i in 2:8)lines(trw[[i]],col=i)
}
