% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{which.mean}
\alias{which.mean}
\title{Where is the mean()}
\usage{
which.mean(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
reference number of vector v
}
\description{
Where is the mean()
}
\examples{
which.mean(1:10)
}
