% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShapeFile.R
\name{WriteShapefile_AnnualRings}
\alias{WriteShapefile_AnnualRings}
\title{WriteShapefile_AnnualRings}
\usage{
WriteShapefile_AnnualRings(L2 = L, filename = "test")
}
\arguments{
\item{L2}{is as list of annual ring polygons (X, Y)}

\item{filename}{is a file name of shape file written to disk.
The extension (.shp) is unnecessary.}
}
\value{
data of Shapefile
}
\description{
WriteShapefile_AnnualRings
}
\examples{
補完推定した年輪のshapefile書き出し ####
WriteShapefile_AnnualRings (L, "test")

}
