% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRingGrowth.R
\name{YearArea}
\alias{YearArea}
\title{relationships year and tree(annual) ring area}
\usage{
YearArea(L2, yr_end = 2018)
}
\arguments{
\item{YA}{}
}
\description{
relationships year and tree(annual) ring area
}
\examples{
L2<-TR@L2
YearArea(L2,yr_end=2018)

}
